整理warning

+++ DingshanSwift/CommonFiles/CocoaSecurity.m
@@ -554,7 +554,7 @@
     if (accumulator > 2) outputLength++;
     
     //truncate data to match actual output length
-    outputData.length = outputLength;
+    outputData.length = (NSUInteger)outputLength;
     return outputLength? outputData: nil;
 }
 - (NSData *)hex:(NSString *)data
